:poop: Renames component with correct spelling
